Switch seed field to biginteger

+++ bingosync-app/bingosync/models/events.py
@@ -75,7 +75,7 @@ class ChatEvent(Event):
 
 class NewCardEvent(Event):
     game_type_value = models.IntegerField(choices=GameType.choices())
-    seed = models.IntegerField(default=0)
+    seed = models.BigIntegerField(default=0)
     hide_card = models.BooleanField(default=False)
 
     @property
